CLIENT + SERVER - handle post and delete on administration page for message

+++ workspaces/server/src/message_error/message.error.service.ts
@@ -1,5 +1,7 @@
+import { UserDocument } from '@app/auth/schemas/user.schema';
 import { MessageErrorDocument } from '@app/message_error/schemas/message.error.schema';
-import { Injectable } from '@nestjs/common';
+import { AuthenticatedRequest } from '@app/types/authenticated.request.type';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 
@@ -10,13 +12,18 @@ export class MessageErrorService {
   constructor(
     @InjectModel('MessageError')
     private messageErrorModel: Model<MessageErrorDocument>,
+    @InjectModel('User') private userModel: Model<UserDocument>,
   ) {}
 
   async getMessageError() {
     return await this.messageErrorModel.findOne();
   }
 
-  async updateMessageError(updateData: EditMessageErrorDto) {
+  async updateMessageError(
+    req: AuthenticatedRequest,
+    updateData: EditMessageErrorDto,
+  ) {
+    await this.checkUserAdmin(req.user.userId);
     const messageError = await this.messageErrorModel.findOne();
 
     if (messageError) {
@@ -28,9 +35,18 @@ export class MessageErrorService {
     return { message: "Le message d'erreur est mis à jour." };
   }
 
-  async deleteMessageError() {
+  async deleteMessageError(req: AuthenticatedRequest) {
+    await this.checkUserAdmin(req.user.userId);
     await this.messageErrorModel.findOneAndDelete();
 
     return { message: "Le message d'erreur a été supprimé." };
+  }
+
+  async checkUserAdmin(userId: string) {
+    const user = await this.userModel.findById(userId).select('isAdmin');
+
+    if (!user || (user && !user?.isAdmin)) {
+      throw new UnauthorizedException('Not an administrator');
+    }
   }
 }
